backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.3
pytest==8.1.1
pytest-django==4.8.0
python-dotenv==1.0.1
Pillow==10.3.0
openpyxl==3.1.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.1.0
